requirements.txt
Flask==2.2.3
tensorflow==2.17.0
Pillow==9.4.0

